refactor(src/model): move schema to model
