+++ src/api/api_od_test.go
@@ -1,20 +1,20 @@
 package api
 
-import (
-	"testing"
-)
+// import (
+// 	"testing"
+// )
 
-func TestOxfordDict(t *testing.T) {
-	word := "word"
-	api := NewOxfordDict()
-	err := api.Query(word, "en")
-	if err != nil {
-		t.Fatal(err)
-	}
-	// assert.NoError(t, err)
+// func TestOxfordDict(t *testing.T) {
+// 	word := "word"
+// 	api := NewOxfordDict()
+// 	err := api.Query(word, "en")
+// 	if err != nil {
+// 		t.Fatal(err)
+// 	}
+// 	assert.NoError(t, err)
 
-	// define, err := api.Parse()
-	// assert.NoError(t, err)
+// 	define, err := api.Parse()
+// 	assert.NoError(t, err)
 
-	// assert.Len(t, define.String(), 1430)
-}
+// 	assert.Len(t, define.String(), 1430)
+// }
